+++ Jam/Assets/Scenes/Case1.unity
@@ -7489,7 +7489,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 2080090357}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 4a81c6d78d5aaaa4b87c91d64536072c, type: 3}
+  m_Script: {fileID: 11500000, guid: d28efdcaa83534044a9be9a46eabee03, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   OnClick:
@@ -7508,6 +7508,8 @@ MonoBehaviour:
           m_BoolArgument: 0
         m_CallState: 2
   yonly: 1
+  minY: -3.73
+  maxY: 3.73
 --- !u!114 &2080090365
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -7731,6 +7733,10 @@ PrefabInstance:
     - target: {fileID: 5402242185366720766, guid: fbbf98b8330b37148a5353e0099d1724, type: 3}
       propertyPath: m_Name
       value: button (2)
+      objectReference: {fileID: 0}
+    - target: {fileID: 5402242185366720766, guid: fbbf98b8330b37148a5353e0099d1724, type: 3}
+      propertyPath: m_IsActive
+      value: 1
       objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: fbbf98b8330b37148a5353e0099d1724, type: 3}
