+++ Jam/Assets/Scripts/ClickandDrag.cs
@@ -2,14 +2,12 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-public class ClikcandDrag : ClickEvent
+public class ClickandDrag : ClickEvent
 {
     bool selected = false;
-    void Start()
-    {
-        
-    }
+    [SerializeField] bool yonly = false;
 
+    [SerializeField] private float minY, maxY;
     
     void Update()
     {
@@ -18,10 +16,37 @@ public class ClikcandDrag : ClickEvent
             selected = false;
             Release();
         }
+        if (selected)
+        {
+            FollowMouse();
+        }
     }
     
     public virtual void Release()
     {
 
+    }
+
+    public void FollowMouse()
+    {
+        Vector3 pos  = Camera.main.ScreenToWorldPoint( new Vector3(Input.mousePosition.x, Input.mousePosition.y, 0));
+        Debug.Log(pos.y);
+        if (yonly)
+        {
+            if (pos.y <= maxY && pos.y >= minY)
+            {
+                transform.position = new Vector3(transform.position.x, pos.y, 0);
+            }
+        }
+        else
+        {
+            transform.position = new Vector3(pos.x, pos.y, 0);
+        }
+        
+    }
+
+    public void OnSelected()
+    {
+        selected = true;
     }
 }
