--- Jam/Assets/Scripts/ClikcandDrag.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d28efdcaa83534044a9be9a46eabee03
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
